Realizados mas repositories y arreglados errores etc. Leer TODOs
Realizados UserRepository.kt y BeerRepository.kt

+++ app/src/main/java/com/unex/asee/ga02/beergo/model/Comment.kt
@@ -40,7 +40,7 @@ data class Comment(
     // Identificador único autoincremental del comentario.
     @PrimaryKey(autoGenerate = true) var commentId: Long,
     // Referencia a la cerveza asociada al comentario.
-    val beerId: Int,
+    val beerId: Long,
     // Usuario que realizó el comentario.
     val userId: Long,
     // El comentario en sí.
